backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
anthropic==0.7.0
python-dotenv==1.0.0
stripe
gunicorn
werkzeug
psycopg2-binary